initial agents baseline #79
