+++ packages/server/src/services/monitoring/switchboard.ts
@@ -1,64 +1,12 @@
 import EventEmitter from 'node:events'
 
-import { Registry } from '@polkadot/types-codec/types'
-import { ControlQuery, extractEvents, extractTxWithEvents, flattenCalls, types } from '@sodazone/ocelloids-sdk'
-import { Observable, filter, from, map, share, switchMap } from 'rxjs'
+import { Logger, Services } from '../types.js'
+import { AgentId, Subscription, SubscriptionStats, XcmEventListener } from './types.js'
 
-import { Logger, NetworkURN, Services } from '../types.js'
-import { extractXcmpReceive, extractXcmpSend } from './ops/xcmp.js'
-import {
-  BridgeSubscription,
-  BridgeType,
-  HexString,
-  RxSubscriptionHandler,
-  RxSubscriptionWithId,
-  Subscription,
-  SubscriptionStats,
-  XcmBridgeAcceptedWithContext,
-  XcmBridgeDeliveredWithContext,
-  XcmBridgeInboundWithContext,
-  XcmEventListener,
-  XcmInbound,
-  XcmInboundWithContext,
-  XcmNotificationType,
-  XcmNotifyMessage,
-  XcmRelayedWithContext,
-  XcmSentWithContext,
-} from './types.js'
-
+import { AgentService } from 'agents/types.js'
 import { NotifierHub } from '../notification/hub.js'
 import { NotifierEvents } from '../notification/types.js'
-import { SubsStore } from '../persistence/subs.js'
 import { TelemetryCollect, TelemetryEventEmitter } from '../telemetry/types.js'
-import { MatchingEngine } from './matching.js'
-
-import { errorMessage } from '../../errors.js'
-import { IngressConsumer } from '../ingress/index.js'
-import { mapXcmSent } from './ops/common.js'
-import { matchMessage, matchSenders, messageCriteria, sendersCriteria } from './ops/criteria.js'
-import { extractDmpReceive, extractDmpSend, extractDmpSendByEvent } from './ops/dmp.js'
-import { extractRelayReceive } from './ops/relay.js'
-import { extractUmpReceive, extractUmpSend } from './ops/ump.js'
-
-import { getChainId, getConsensus } from '../config.js'
-import { extractBridgeMessageAccepted, extractBridgeMessageDelivered, extractBridgeReceive } from './ops/pk-bridge.js'
-import { getBridgeHubNetworkId } from './ops/util.js'
-import {
-  dmpDownwardMessageQueuesKey,
-  parachainSystemHrmpOutboundMessages,
-  parachainSystemUpwardMessages,
-} from './storage.js'
-import {
-  GetDownwardMessageQueues,
-  GetOutboundHrmpMessages,
-  GetOutboundUmpMessages,
-  GetStorageAt,
-} from './types-augmented.js'
-
-type Monitor = {
-  subs: RxSubscriptionWithId[]
-  controls: Record<string, ControlQuery>
-}
 
 export enum SubscribeErrorCodes {
   TOO_MANY_SUBSCRIBERS,
@@ -80,8 +28,6 @@ export type SwitchboardOptions = {
   subscriptionMaxEphemeral?: number
 }
 
-const SUB_ERROR_RETRY_MS = 5000
-
 /**
  * XCM Subscriptions Switchboard.
  *
@@ -93,31 +39,19 @@ const SUB_ERROR_RETRY_MS = 5000
  */
 export class Switchboard extends (EventEmitter as new () => TelemetryEventEmitter) {
   readonly #log: Logger
-  readonly #db: SubsStore
-  readonly #engine: MatchingEngine
-  readonly #ingress: IngressConsumer
   readonly #notifier: NotifierHub
   readonly #stats: SubscriptionStats
   readonly #maxEphemeral: number
   readonly #maxPersistent: number
-  readonly #timeouts: NodeJS.Timeout[] = []
-
-  #subs: Record<string, RxSubscriptionHandler> = {}
-  #shared: {
-    blockEvents: Record<string, Observable<types.BlockEvent>>
-    blockExtrinsics: Record<string, Observable<types.TxWithIdAndEvent>>
-  }
+  readonly #agentService: AgentService
 
   constructor(ctx: Services, options: SwitchboardOptions) {
     super()
 
-    const { log, subsStore, ingressConsumer } = ctx
+    this.#log = ctx.log
+    this.#agentService = ctx.agentService
 
-    this.#db = subsStore
-    this.#log = log
-
-    this.#engine = new MatchingEngine(ctx, this.#onXcmWaypointReached.bind(this))
-    this.#ingress = ingressConsumer
+    // TODO here could be local for websockets over event emitter or remote over redis
     this.#notifier = new NotifierHub(ctx)
     this.#stats = {
       ephemeral: 0,
@@ -125,29 +59,29 @@ export class Switchboard extends (EventEmitter as new () => TelemetryEventEmitte
     }
     this.#maxEphemeral = options.subscriptionMaxEphemeral ?? 10_000
     this.#maxPersistent = options.subscriptionMaxPersistent ?? 10_000
-    this.#shared = {
-      blockEvents: {},
-      blockExtrinsics: {},
-    }
   }
 
   /**
-   * Subscribes according to the given query subscription.
+   * Subscribes according to the given subscription.
    *
-   * @param {Subscription} qs The query subscription.
+   * @param {Subscription} s The subscription.
    * @throws {SubscribeError} If there is an error creating the subscription.
    */
-  async subscribe(qs: Subscription) {
+  async subscribe(s: Subscription) {
     if (this.#stats.ephemeral >= this.#maxEphemeral || this.#stats.persistent >= this.#maxPersistent) {
       throw new SubscribeError(SubscribeErrorCodes.TOO_MANY_SUBSCRIBERS, 'too many subscriptions')
     }
 
-    if (!qs.ephemeral) {
-      await this.#db.insert(qs)
-    }
-    this.#monitor(qs)
+    const agent = this.#agentService.getAgentById(s.agent)
+    await agent.subscribe(s)
 
-    this.#log.info('[%s] new subscription: %j', qs.origin, qs)
+    this.#log.info('[%s] new subscription: %j', s.agent, s)
+
+    if (s.ephemeral) {
+      this.#stats.ephemeral++
+    } else {
+      this.#stats.persistent++
+    }
   }
 
   /**
@@ -171,144 +105,41 @@ export class Switchboard extends (EventEmitter as new () => TelemetryEventEmitte
   }
 
   /**
-   * Unsubscribes by subsciption identifier.
+   * Unsubscribes by subsciption and agent identifier.
    *
    * If the subscription does not exists just ignores it.
    *
+   * @param {AgentId} agentId The agent identifier.
    * @param {string} id The subscription identifier.
    */
-  async unsubscribe(id: string) {
-    if (this.#subs[id] === undefined) {
-      this.#log.warn('unsubscribe from a non-existent subscription %s', id)
-      return
-    }
+  async unsubscribe(agentId: AgentId, id: string) {
+    const agent = this.#agentService.getAgentById(agentId)
+    const { ephemeral } = agent.getSubscriptionHandler(id)
+    await agent.unsubscribe(id)
 
-    try {
-      const {
-        descriptor: { origin, ephemeral },
-        originSubs,
-        destinationSubs,
-        relaySub,
-      } = this.#subs[id]
-
-      this.#log.info('[%s] unsubscribe %s', origin, id)
-
-      originSubs.forEach(({ sub }) => sub.unsubscribe())
-      destinationSubs.forEach(({ sub }) => sub.unsubscribe())
-      if (relaySub) {
-        relaySub.sub.unsubscribe()
-      }
-      delete this.#subs[id]
-
-      await this.#engine.clearPendingStates(id)
-
-      if (ephemeral) {
-        this.#stats.ephemeral--
-      } else {
-        this.#stats.persistent--
-        await this.#db.remove(id)
-      }
-    } catch (error) {
-      this.#log.error(error, 'Error unsubscribing %s', id)
+    if (ephemeral) {
+      this.#stats.ephemeral--
+    } else {
+      this.#stats.persistent--
     }
   }
 
   async start() {
-    await this.#startNetworkMonitors()
+    // empty
   }
 
   /**
-   * Stops the switchboard and unsubscribes from the underlying
-   * reactive subscriptions.
+   * Stops the switchboard.
    */
   async stop() {
-    this.#log.info('Stopping switchboard')
-
-    for (const {
-      descriptor: { id },
-      originSubs,
-      destinationSubs,
-      relaySub,
-    } of Object.values(this.#subs)) {
-      this.#log.info('Unsubscribe %s', id)
-
-      originSubs.forEach(({ sub }) => sub.unsubscribe())
-      destinationSubs.forEach(({ sub }) => sub.unsubscribe())
-      if (relaySub) {
-        relaySub.sub.unsubscribe()
-      }
-    }
-
-    for (const t of this.#timeouts) {
-      t.unref()
-    }
-
-    await this.#engine.stop()
+    // empty
   }
 
   /**
    * Gets a subscription handler by id.
    */
-  findSubscriptionHandler(id: string) {
-    return this.#subs[id]
-  }
-
-  /**
-   * Updates the senders control handler.
-   *
-   * Applies to the outbound extrinsic signers.
-   */
-  updateSenders(id: string) {
-    const {
-      descriptor: { senders },
-      sendersControl,
-    } = this.#subs[id]
-
-    sendersControl.change(sendersCriteria(senders))
-  }
-
-  /**
-   * Updates the message control handler.
-   *
-   * Updates the destination subscriptions.
-   */
-  updateDestinations(id: string) {
-    const { descriptor, messageControl } = this.#subs[id]
-
-    messageControl.change(messageCriteria(descriptor.destinations as NetworkURN[]))
-
-    const updatedSubs = this.#updateDestinationSubscriptions(id)
-    this.#subs[id].destinationSubs = updatedSubs
-  }
-
-  /**
-   * Updates the subscription to relayed HRMP messages in the relay chain.
-   */
-  updateEvents(id: string) {
-    const { descriptor, relaySub } = this.#subs[id]
-
-    if (this.#shouldMonitorRelay(descriptor) && relaySub === undefined) {
-      try {
-        this.#subs[id].relaySub = this.#monitorRelay(descriptor)
-      } catch (error) {
-        // log instead of throw to not block OD subscriptions
-        this.#log.error(error, 'Error on relay subscription (%s)', id)
-      }
-    } else if (!this.#shouldMonitorRelay(descriptor) && relaySub !== undefined) {
-      relaySub.sub.unsubscribe()
-      delete this.#subs[id].relaySub
-    }
-  }
-
-  /**
-   * Updates a subscription descriptor.
-   */
-  updateSubscription(sub: Subscription) {
-    if (this.#subs[sub.id]) {
-      this.#subs[sub.id].descriptor = sub
-    } else {
-      this.#log.warn('trying to update an unknown subscription %s', sub.id)
-    }
+  findSubscriptionHandler(agentId: AgentId, id: string) {
+    return this.#agentService.getAgentById(agentId).getSubscriptionHandler(id)
   }
 
   /**
@@ -318,7 +149,7 @@ export class Switchboard extends (EventEmitter as new () => TelemetryEventEmitte
    */
   collectTelemetry(collect: TelemetryCollect) {
     collect(this)
-    collect(this.#engine)
+    // collect(this.#engine)
     collect(this.#notifier)
   }
 
@@ -327,639 +158,5 @@ export class Switchboard extends (EventEmitter as new () => TelemetryEventEmitte
    */
   get stats() {
     return this.#stats
-  }
-
-  /**
-   * Main monitoring logic.
-   *
-   * This method sets up and manages subscriptions for XCM messages based on the provided
-   * query subscription information. It creates subscriptions for both the origin and destination
-   * networks, monitors XCM message transfers, and emits events accordingly.
-   *
-   * @param {Subscription} qs - The query subscription.
-   * @throws {Error} If there is an error during the subscription setup process.
-   * @private
-   */
-  #monitor(qs: Subscription) {
-    const { id } = qs
-
-    let origMonitor: Monitor = { subs: [], controls: {} }
-    let destMonitor: Monitor = { subs: [], controls: {} }
-    const bridgeSubs: BridgeSubscription[] = []
-    let relaySub: RxSubscriptionWithId | undefined
-
-    try {
-      origMonitor = this.#monitorOrigins(qs)
-      destMonitor = this.#monitorDestinations(qs)
-    } catch (error) {
-      // Clean up origin subscriptions.
-      origMonitor.subs.forEach(({ sub }) => {
-        sub.unsubscribe()
-      })
-      throw error
-    }
-
-    // Only subscribe to relay events if required by subscription.
-    // Contained in its own try-catch so it doesn't prevent origin-destination subs in case of error.
-    if (this.#shouldMonitorRelay(qs)) {
-      try {
-        relaySub = this.#monitorRelay(qs)
-      } catch (error) {
-        // log instead of throw to not block OD subscriptions
-        this.#log.error(error, 'Error on relay subscription (%s)', id)
-      }
-    }
-
-    if (qs.bridges !== undefined) {
-      if (qs.bridges.includes('pk-bridge')) {
-        try {
-          bridgeSubs.push(this.#monitorPkBridge(qs))
-        } catch (error) {
-          // log instead of throw to not block OD subscriptions
-          this.#log.error(error, 'Error on bridge subscription (%s)', id)
-        }
-      }
-    }
-
-    if (qs.bridges !== undefined) {
-      if (qs.bridges.includes('pk-bridge')) {
-        try {
-          bridgeSubs.push(this.#monitorPkBridge(qs))
-        } catch (error) {
-          // log instead of throw to not block OD subscriptions
-          this.#log.error(error, 'Error on bridge subscription (%s)', id)
-        }
-      }
-    }
-
-    const { sendersControl, messageControl } = origMonitor.controls
-
-    this.#subs[id] = {
-      descriptor: qs,
-      sendersControl,
-      messageControl,
-      originSubs: origMonitor.subs,
-      destinationSubs: destMonitor.subs,
-      bridgeSubs,
-      relaySub,
-    }
-
-    if (qs.ephemeral) {
-      this.#stats.ephemeral++
-    } else {
-      this.#stats.persistent++
-    }
-  }
-
-  /**
-   * Set up inbound monitors for XCM protocols.
-   *
-   * @private
-   */
-  #monitorDestinations({ id, destinations, origin }: Subscription): Monitor {
-    const subs: RxSubscriptionWithId[] = []
-    const originId = origin as NetworkURN
-    try {
-      for (const dest of destinations as NetworkURN[]) {
-        const chainId = dest
-        if (this.#subs[id]?.destinationSubs.find((s) => s.chainId === chainId)) {
-          // Skip existing subscriptions
-          // for the same destination chain
-          continue
-        }
-
-        const inboundObserver = {
-          error: (error: any) => {
-            this.#log.error(error, '[%s] error on destination subscription %s', chainId, id)
-            this.emit('telemetrySubscriptionError', {
-              subscriptionId: id,
-              chainId,
-              direction: 'in',
-            })
-
-            // try recover inbound subscription
-            if (this.#subs[id]) {
-              const { destinationSubs } = this.#subs[id]
-              const index = destinationSubs.findIndex((s) => s.chainId === chainId)
-              if (index > -1) {
-                destinationSubs.splice(index, 1)
-                this.#timeouts.push(
-                  setTimeout(() => {
-                    this.#log.info(
-                      '[%s] UPDATE destination subscription %s due error %s',
-                      chainId,
-                      id,
-                      errorMessage(error)
-                    )
-                    const updated = this.#updateDestinationSubscriptions(id)
-                    this.#subs[id].destinationSubs = updated
-                  }, SUB_ERROR_RETRY_MS)
-                )
-              }
-            }
-          },
-        }
-
-        if (this.#ingress.isRelay(dest)) {
-          // VMP UMP
-          this.#log.info('[%s] subscribe inbound UMP (%s)', chainId, id)
-
-          subs.push({
-            chainId,
-            sub: this.#sharedBlockEvents(chainId)
-              .pipe(extractUmpReceive(originId), this.#emitInbound(id, chainId))
-              .subscribe(inboundObserver),
-          })
-        } else if (this.#ingress.isRelay(originId)) {
-          // VMP DMP
-          this.#log.info('[%s] subscribe inbound DMP (%s)', chainId, id)
-
-          subs.push({
-            chainId,
-            sub: this.#sharedBlockEvents(chainId)
-              .pipe(extractDmpReceive(), this.#emitInbound(id, chainId))
-              .subscribe(inboundObserver),
-          })
-        } else {
-          // Inbound HRMP / XCMP transport
-          this.#log.info('[%s] subscribe inbound HRMP (%s)', chainId, id)
-
-          subs.push({
-            chainId,
-            sub: this.#sharedBlockEvents(chainId)
-              .pipe(extractXcmpReceive(), this.#emitInbound(id, chainId))
-              .subscribe(inboundObserver),
-          })
-        }
-      }
-    } catch (error) {
-      // Clean up subscriptions.
-      subs.forEach(({ sub }) => {
-        sub.unsubscribe()
-      })
-      throw error
-    }
-
-    return { subs, controls: {} }
-  }
-
-  /**
-   * Set up outbound monitors for XCM protocols.
-   *
-   * @private
-   */
-  #monitorOrigins({ id, origin, senders, destinations }: Subscription): Monitor {
-    const subs: RxSubscriptionWithId[] = []
-    const chainId = origin as NetworkURN
-
-    if (this.#subs[id]?.originSubs.find((s) => s.chainId === chainId)) {
-      throw new Error(`Fatal: duplicated origin monitor ${id} for chain ${chainId}`)
-    }
-
-    const sendersControl = ControlQuery.from(sendersCriteria(senders))
-    const messageControl = ControlQuery.from(messageCriteria(destinations as NetworkURN[]))
-
-    const outboundObserver = {
-      error: (error: any) => {
-        this.#log.error(error, '[%s] error on origin subscription %s', chainId, id)
-        this.emit('telemetrySubscriptionError', {
-          subscriptionId: id,
-          chainId,
-          direction: 'out',
-        })
-
-        // try recover outbound subscription
-        // note: there is a single origin per outbound
-        if (this.#subs[id]) {
-          const { originSubs, descriptor } = this.#subs[id]
-          const index = originSubs.findIndex((s) => s.chainId === chainId)
-          if (index > -1) {
-            this.#subs[id].originSubs = []
-            this.#timeouts.push(
-              setTimeout(() => {
-                if (this.#subs[id]) {
-                  this.#log.info('[%s] UPDATE origin subscription %s due error %s', chainId, id, errorMessage(error))
-                  const { subs: updated, controls } = this.#monitorOrigins(descriptor)
-                  this.#subs[id].sendersControl = controls.sendersControl
-                  this.#subs[id].messageControl = controls.messageControl
-                  this.#subs[id].originSubs = updated
-                }
-              }, SUB_ERROR_RETRY_MS)
-            )
-          }
-        }
-      },
-    }
-
-    try {
-      if (this.#ingress.isRelay(chainId)) {
-        // VMP DMP
-        this.#log.info('[%s] subscribe outbound DMP (%s)', chainId, id)
-
-        subs.push({
-          chainId,
-          sub: this.#ingress
-            .getRegistry(chainId)
-            .pipe(
-              switchMap((registry) =>
-                this.#sharedBlockExtrinsics(chainId).pipe(
-                  extractDmpSend(chainId, this.#getDmp(chainId, registry), registry),
-                  this.#emitOutbound(id, chainId, registry, messageControl)
-                )
-              )
-            )
-            .subscribe(outboundObserver),
-        })
-
-        // VMP DMP
-        this.#log.info('[%s] subscribe outbound DMP - by event (%s)', chainId, id)
-
-        subs.push({
-          chainId,
-          sub: this.#ingress
-            .getRegistry(chainId)
-            .pipe(
-              switchMap((registry) =>
-                this.#sharedBlockEvents(chainId).pipe(
-                  extractDmpSendByEvent(chainId, this.#getDmp(chainId, registry), registry),
-                  this.#emitOutbound(id, chainId, registry, messageControl)
-                )
-              )
-            )
-            .subscribe(outboundObserver),
-        })
-      } else {
-        // Outbound HRMP / XCMP transport
-        this.#log.info('[%s] subscribe outbound HRMP (%s)', chainId, id)
-
-        subs.push({
-          chainId,
-          sub: this.#ingress
-            .getRegistry(chainId)
-            .pipe(
-              switchMap((registry) =>
-                this.#sharedBlockEvents(chainId).pipe(
-                  extractXcmpSend(chainId, this.#getHrmp(chainId, registry), registry),
-                  this.#emitOutbound(id, chainId, registry, messageControl)
-                )
-              )
-            )
-            .subscribe(outboundObserver),
-        })
-
-        // VMP UMP
-        this.#log.info('[%s] subscribe outbound UMP (%s)', chainId, id)
-
-        subs.push({
-          chainId,
-          sub: this.#ingress
-            .getRegistry(chainId)
-            .pipe(
-              switchMap((registry) =>
-                this.#sharedBlockEvents(chainId).pipe(
-                  extractUmpSend(chainId, this.#getUmp(chainId, registry), registry),
-                  this.#emitOutbound(id, chainId, registry, messageControl)
-                )
-              )
-            )
-            .subscribe(outboundObserver),
-        })
-      }
-    } catch (error) {
-      // Clean up subscriptions.
-      subs.forEach(({ sub }) => {
-        sub.unsubscribe()
-      })
-      throw error
-    }
-
-    return {
-      subs,
-      controls: {
-        sendersControl,
-        messageControl,
-      },
-    }
-  }
-
-  #monitorRelay({ id, destinations, origin }: Subscription) {
-    const chainId = origin as NetworkURN
-    if (this.#subs[id]?.relaySub) {
-      this.#log.debug('Relay subscription already exists.')
-    }
-    const messageControl = ControlQuery.from(messageCriteria(destinations as NetworkURN[]))
-
-    const emitRelayInbound = () => (source: Observable<XcmRelayedWithContext>) =>
-      source.pipe(switchMap((message) => from(this.#engine.onRelayedMessage(id, message))))
-
-    const relayObserver = {
-      error: (error: any) => {
-        this.#log.error(error, '[%s] error on relay subscription s', chainId, id)
-        this.emit('telemetrySubscriptionError', {
-          subscriptionId: id,
-          chainId,
-          direction: 'relay',
-        })
-
-        // try recover relay subscription
-        // there is only one subscription per subscription ID for relay
-        if (this.#subs[id]) {
-          this.#timeouts.push(
-            setTimeout(async () => {
-              this.#log.info('[%s] UPDATE relay subscription %s due error %s', chainId, id, errorMessage(error))
-              const updatedSub = await this.#monitorRelay(this.#subs[id].descriptor)
-              this.#subs[id].relaySub = updatedSub
-            }, SUB_ERROR_RETRY_MS)
-          )
-        }
-      },
-    }
-
-    // TODO: should resolve relay id for consensus in context
-    const relayIds = this.#ingress.getRelayIds()
-    const relayId = relayIds.find((r) => getConsensus(r) === getConsensus(chainId))
-
-    if (relayId === undefined) {
-      throw new Error(`No relay ID found for chain ${chainId}`)
-    }
-    this.#log.info('[%s] subscribe relay %s xcm events (%s)', chainId, relayId, id)
-    return {
-      chainId,
-      sub: this.#ingress
-        .getRegistry(relayId)
-        .pipe(
-          switchMap((registry) =>
-            this.#sharedBlockExtrinsics(relayId).pipe(
-              extractRelayReceive(chainId, messageControl, registry),
-              emitRelayInbound()
-            )
-          )
-        )
-        .subscribe(relayObserver),
-    }
-  }
-
-  // Assumes only 1 pair of bridge hub origin-destination is possible
-  // TODO: handle possible multiple different consensus utilizing PK bridge e.g. solochains?
-  #monitorPkBridge({ id, destinations, origin }: Subscription) {
-    const originBridgeHub = getBridgeHubNetworkId(origin as NetworkURN)
-    const dest = (destinations as NetworkURN[]).find((d) => getConsensus(d) !== getConsensus(origin as NetworkURN))
-
-    if (dest === undefined) {
-      throw new Error(`No destination on different consensus found for bridging (sub=${id})`)
-    }
-
-    const destBridgeHub = getBridgeHubNetworkId(dest)
-
-    if (originBridgeHub === undefined || destBridgeHub === undefined) {
-      throw new Error(
-        `Unable to subscribe to PK bridge due to missing bridge hub network URNs for origin=${origin} and destinations=${destinations}. (sub=${id})`
-      )
-    }
-
-    if (this.#subs[id]?.bridgeSubs.find((s) => s.type === 'pk-bridge')) {
-      throw new Error(`Fatal: duplicated PK bridge monitor ${id}`)
-    }
-
-    const type: BridgeType = 'pk-bridge'
-
-    const emitBridgeOutboundAccepted = () => (source: Observable<XcmBridgeAcceptedWithContext>) =>
-      source.pipe(switchMap((message) => from(this.#engine.onBridgeOutboundAccepted(id, message))))
-
-    const emitBridgeOutboundDelivered = () => (source: Observable<XcmBridgeDeliveredWithContext>) =>
-      source.pipe(switchMap((message) => from(this.#engine.onBridgeOutboundDelivered(id, message))))
-
-    const emitBridgeInbound = () => (source: Observable<XcmBridgeInboundWithContext>) =>
-      source.pipe(switchMap((message) => from(this.#engine.onBridgeInbound(id, message))))
-
-    const pkBridgeObserver = {
-      error: (error: any) => {
-        this.#log.error(error, '[%s] error on PK bridge subscription s', originBridgeHub, id)
-        // this.emit('telemetrySubscriptionError', {
-        //   subscriptionId: id,
-        //   chainId: originBridgeHub,
-        //   direction: 'bridge',
-        // });
-
-        // try recover pk bridge subscription
-        if (this.#subs[id]) {
-          const { bridgeSubs } = this.#subs[id]
-          const index = bridgeSubs.findIndex((s) => s.type === 'pk-bridge')
-          if (index > -1) {
-            bridgeSubs.splice(index, 1)
-            this.#timeouts.push(
-              setTimeout(() => {
-                this.#log.info(
-                  '[%s] UPDATE destination subscription %s due error %s',
-                  originBridgeHub,
-                  id,
-                  errorMessage(error)
-                )
-                bridgeSubs.push(this.#monitorPkBridge(this.#subs[id].descriptor))
-                this.#subs[id].bridgeSubs = bridgeSubs
-              }, SUB_ERROR_RETRY_MS)
-            )
-          }
-        }
-      },
-    }
-
-    this.#log.info(
-      '[%s] subscribe PK bridge outbound accepted events on bridge hub %s (%s)',
-      origin,
-      originBridgeHub,
-      id
-    )
-    const outboundAccepted: RxSubscriptionWithId = {
-      chainId: originBridgeHub,
-      sub: this.#ingress
-        .getRegistry(originBridgeHub)
-        .pipe(
-          switchMap((registry) =>
-            this.#sharedBlockEvents(originBridgeHub).pipe(
-              extractBridgeMessageAccepted(originBridgeHub, registry, this.#getStorageAt(originBridgeHub)),
-              emitBridgeOutboundAccepted()
-            )
-          )
-        )
-        .subscribe(pkBridgeObserver),
-    }
-
-    this.#log.info(
-      '[%s] subscribe PK bridge outbound delivered events on bridge hub %s (%s)',
-      origin,
-      originBridgeHub,
-      id
-    )
-    const outboundDelivered: RxSubscriptionWithId = {
-      chainId: originBridgeHub,
-      sub: this.#ingress
-        .getRegistry(originBridgeHub)
-        .pipe(
-          switchMap((registry) =>
-            this.#sharedBlockEvents(originBridgeHub).pipe(
-              extractBridgeMessageDelivered(originBridgeHub, registry),
-              emitBridgeOutboundDelivered()
-            )
-          )
-        )
-        .subscribe(pkBridgeObserver),
-    }
-
-    this.#log.info('[%s] subscribe PK bridge inbound events on bridge hub %s (%s)', origin, destBridgeHub, id)
-    const inbound: RxSubscriptionWithId = {
-      chainId: destBridgeHub,
-      sub: this.#sharedBlockEvents(destBridgeHub)
-        .pipe(extractBridgeReceive(destBridgeHub), emitBridgeInbound())
-        .subscribe(pkBridgeObserver),
-    }
-
-    return {
-      type,
-      subs: [outboundAccepted, outboundDelivered, inbound],
-    }
-  }
-
-  #updateDestinationSubscriptions(id: string) {
-    const { descriptor, destinationSubs } = this.#subs[id]
-    // Subscribe to new destinations, if any
-    const { subs } = this.#monitorDestinations(descriptor)
-    const updatedSubs = destinationSubs.concat(subs)
-    // Unsubscribe removed destinations, if any
-    const removed = updatedSubs.filter((s) => !descriptor.destinations.includes(s.chainId))
-    removed.forEach(({ sub }) => sub.unsubscribe())
-    // Return list of updated subscriptions
-    return updatedSubs.filter((s) => !removed.includes(s))
-  }
-
-  /**
-   * Starts collecting XCM messages.
-   *
-   * Monitors all the active subscriptions for the configured networks.
-   *
-   * @private
-   */
-  async #startNetworkMonitors() {
-    const chainIds = this.#ingress.getChainIds()
-
-    for (const chainId of chainIds) {
-      const subs = await this.#db.getByNetworkId(chainId)
-
-      this.#log.info('[%s] #subscriptions %d', chainId, subs.length)
-
-      for (const sub of subs) {
-        try {
-          this.#monitor(sub)
-        } catch (err) {
-          this.#log.error(err, 'Unable to create subscription: %j', sub)
-        }
-      }
-    }
-  }
-
-  #onXcmWaypointReached(msg: XcmNotifyMessage) {
-    const { subscriptionId } = msg
-    if (this.#subs[subscriptionId]) {
-      const { descriptor, sendersControl } = this.#subs[subscriptionId]
-      if (
-        (descriptor.events === undefined || descriptor.events === '*' || descriptor.events.includes(msg.type)) &&
-        matchSenders(sendersControl, msg.sender)
-      ) {
-        this.#notifier.notify(descriptor, msg)
-      }
-    } else {
-      // this could happen with closed ephemeral subscriptions
-      this.#log.warn('Unable to find descriptor for subscription %s', subscriptionId)
-    }
-  }
-
-  #sharedBlockEvents(chainId: NetworkURN): Observable<types.BlockEvent> {
-    if (!this.#shared.blockEvents[chainId]) {
-      this.#shared.blockEvents[chainId] = this.#ingress.finalizedBlocks(chainId).pipe(extractEvents(), share())
-    }
-    return this.#shared.blockEvents[chainId]
-  }
-
-  #sharedBlockExtrinsics(chainId: NetworkURN): Observable<types.TxWithIdAndEvent> {
-    if (!this.#shared.blockExtrinsics[chainId]) {
-      this.#shared.blockExtrinsics[chainId] = this.#ingress
-        .finalizedBlocks(chainId)
-        .pipe(extractTxWithEvents(), flattenCalls(), share())
-    }
-    return this.#shared.blockExtrinsics[chainId]
-  }
-
-  /**
-   * Checks if relayed HRMP messages should be monitored.
-   *
-   * All of the following conditions needs to be met:
-   * 1. `xcm.relayed` notification event is requested in the subscription
-   * 2. Origin chain is not a relay chain
-   * 3. At least one destination chain is a parachain
-   *
-   * @param Subscription
-   * @returns boolean
-   */
-  #shouldMonitorRelay({ origin, destinations, events }: Subscription) {
-    return (
-      (events === undefined || events === '*' || events.includes(XcmNotificationType.Relayed)) &&
-      !this.#ingress.isRelay(origin as NetworkURN) &&
-      destinations.some((d) => !this.#ingress.isRelay(d as NetworkURN))
-    )
-  }
-
-  #emitInbound(id: string, chainId: NetworkURN) {
-    return (source: Observable<XcmInboundWithContext>) =>
-      source.pipe(switchMap((msg) => from(this.#engine.onInboundMessage(new XcmInbound(id, chainId, msg)))))
-  }
-
-  #emitOutbound(id: string, origin: NetworkURN, registry: Registry, messageControl: ControlQuery) {
-    const {
-      descriptor: { outboundTTL },
-    } = this.#subs[id]
-
-    return (source: Observable<XcmSentWithContext>) =>
-      source.pipe(
-        mapXcmSent(id, registry, origin),
-        filter((msg) => matchMessage(messageControl, msg)),
-        switchMap((outbound) => from(this.#engine.onOutboundMessage(outbound, outboundTTL)))
-      )
-  }
-
-  #getDmp(chainId: NetworkURN, registry: Registry): GetDownwardMessageQueues {
-    return (blockHash: HexString, networkId: NetworkURN) => {
-      const paraId = getChainId(networkId)
-      return from(this.#ingress.getStorage(chainId, dmpDownwardMessageQueuesKey(registry, paraId), blockHash)).pipe(
-        map((buffer) => {
-          return registry.createType('Vec<PolkadotCorePrimitivesInboundDownwardMessage>', buffer)
-        })
-      )
-    }
-  }
-
-  #getUmp(chainId: NetworkURN, registry: Registry): GetOutboundUmpMessages {
-    return (blockHash: HexString) => {
-      return from(this.#ingress.getStorage(chainId, parachainSystemUpwardMessages, blockHash)).pipe(
-        map((buffer) => {
-          return registry.createType('Vec<Bytes>', buffer)
-        })
-      )
-    }
-  }
-
-  #getHrmp(chainId: NetworkURN, registry: Registry): GetOutboundHrmpMessages {
-    return (blockHash: HexString) => {
-      return from(this.#ingress.getStorage(chainId, parachainSystemHrmpOutboundMessages, blockHash)).pipe(
-        map((buffer) => {
-          return registry.createType('Vec<PolkadotCorePrimitivesOutboundHrmpMessage>', buffer)
-        })
-      )
-    }
-  }
-
-  #getStorageAt(chainId: NetworkURN): GetStorageAt {
-    return (blockHash: HexString, key: HexString) => {
-      return from(this.#ingress.getStorage(chainId, key, blockHash))
-    }
   }
 }
