+++ packages/server/src/services/notification/webhook.ts
@@ -3,8 +3,9 @@ import { EventEmitter } from 'node:events'
 import got from 'got'
 import { ulid } from 'ulidx'
 
+import { XcmNotifyMessage } from 'agents/xcm/types.js'
 import version from '../../version.js'
-import { Subscription, WebhookNotification, XcmNotifyMessage } from '../monitoring/types.js'
+import { Subscription, WebhookNotification } from '../monitoring/types.js'
 import { Logger, Services } from '../types.js'
 
 import { Scheduled, Scheduler, SubsStore } from '../persistence/index.js'
@@ -81,9 +82,7 @@ export class WebhookNotifier extends (EventEmitter as new () => NotifierEmitter)
       for (const chan of channels) {
         if (chan.type === 'webhook') {
           const config = chan as WebhookNotification
-          if (config.events === undefined || config.events === '*' || config.events.includes(scheduled.task.msg.type)) {
-            await this.#post(scheduled, config)
-          }
+          await this.#post(scheduled, config)
         }
       }
     } catch (error) {
